feat(219): unify shared/components bounds utilities with @debrief/utils
- Delete shared/components/src/utils/bounds.ts (215 LOC, 6 helpers)
- Port 5 helpers (expandBounds, isPointInBounds, bboxOverlapsViewport,
  viewportToBounds, filterBySpatialExtent) into shared/utils/src/bounds.ts
- Add pre-computed bbox fast-path to calculateBounds (absorbs backlog #211)
  using isValidBboxTuple narrowing predicate — zero any/as-casts (Article XV)
- Add ViewportPolygon type import (dep already declared)
- Extend BoundsInputFeature.bbox to readonly number[] | null | undefined
  so DebriefFeature.bbox (number[]) assigns structurally without casts
- Update shared/utils/src/index.ts barrel with 5 new exports
- Migrate MapView.tsx, LeafletToolbar.tsx, useBrowserFilter.ts to @debrief/utils
- Update shared/components barrel re-exports from @debrief/utils
- Migrate bounds tests from shared/components into shared/utils/tests/bounds.test.ts
  (75 tests; net gain 17 assertions; zero net loss — SC-003)
- Add bounds.types.test-d.ts: compile-time type assertions for FR-016
- Capture evidence: test-summary, usage-example, before-after,
  consolidation-metrics, dep-graph-check
- Add shipped-post.md and linkedin-shipped.md media content
- Mark BACKLOG.md item 213 as implementing

SC-001/SC-002: exactly one calculateBounds implementation in source.
SC-006: zero external consumer import changes.
All 300 @debrief/utils tests pass; all 1647 @debrief/components tests pass.
