+++ BACKLOG.md
@@ -168,7 +168,7 @@ Description formats:
 | 216 | Feature | [[E024 2/4] Storyboarding: capture](specs/216-storyboarding-capture/spec.md) — Ctrl/Cmd+Alt+C shortcut scoped to Map Viewer, first-capture inline quick-pick for Storyboard name, synchronous #174 thumbnail, DTG-default Scene title, duplicate-timestamp Replace / Offset (+1s) / Cancel prompt; minimal panel auto-opens to confirm persistence. First user-visible slice. (requires #215, #174) | 4 | 3 | 4 | 11 | Medium | specified |
 | ~~215~~ | ~~Infrastructure~~ | ~~[[E024 1/4] Storyboarding: schema + CRUD core](specs/215-storyboarding-schema/spec.md) — LinkML `Storyboard` + `Scene` + `Viewport` + `HistoryEntry` sources, generated Pydantic / JSON Schema / TypeScript bindings, Article II adherence tests (golden fixtures, Python↔TS round-trip, Pydantic-vs-LinkML JSON Schema equality), and a headless shared TypeScript CRUD module at `shared/components/storyboard/` enforcing ordering / duplicate-timestamp rejection / `feature_set_hash` / missing-data detection / append-only provenance. Authoritative Key Entities for the whole epic.~~ | ~~3~~ | ~~2~~ | ~~4~~ | ~~9~~ | ~~Medium~~ | ~~complete~~ |
 | ~~214~~ | ~~Tech Debt~~ | ~~[Drift-prevention rule for `@debrief/utils` re-duplication — add a lint or CI check that fails if a file matching `apps/*/src/utils/bounds.ts` reappears, or more generally if any `apps/*` file exports a symbol already exported from `@debrief/utils`. Makes the SC-001 guarantee from #200 durable; without it, a future contributor could reintroduce a local `bounds.ts` without friction. Candidate implementations: custom ESLint rule (`no-restricted-exports`-flavoured), or a small CI script under `scripts/`. (follow-up to #200)](specs/214-utils-drift-guard/spec.md)~~ | ~~3~~ | ~~1~~ | ~~4~~ | ~~8~~ | ~~Low~~ | ~~complete~~ |
-| 213 | Tech Debt | Unify `shared/components/src/utils/bounds.ts` with `@debrief/utils` — that copy operates on LinkML-typed `DebriefFeature` arrays and carries additional helpers (`expandBounds`, `isPointInBounds`, `bboxOverlapsViewport`, `viewportToBounds`, `filterBySpatialExtent`). Unification requires reconciling three feature-type families (`DebriefFeature` / `SafeFeature` / `GeoJSONFeature`) and migrating four consumers. Must also settle the pre-computed-`bbox` fast-path policy on the unified implementation (absorb #211 at approval time if #211 is still open). Legitimately larger scope than #200's generic-GeoJSON consolidation — treat as its own cleanup. (follow-up to #200; absorbs #211 on approval) | 4 | 2 | 3 | 9 | Medium | approved |
+| 213 | Tech Debt | Unify `shared/components/src/utils/bounds.ts` with `@debrief/utils` — that copy operates on LinkML-typed `DebriefFeature` arrays and carries additional helpers (`expandBounds`, `isPointInBounds`, `bboxOverlapsViewport`, `viewportToBounds`, `filterBySpatialExtent`). Unification requires reconciling three feature-type families (`DebriefFeature` / `SafeFeature` / `GeoJSONFeature`) and migrating four consumers. Must also settle the pre-computed-`bbox` fast-path policy on the unified implementation (absorb #211 at approval time if #211 is still open). Legitimately larger scope than #200's generic-GeoJSON consolidation — treat as its own cleanup. (follow-up to #200; absorbs #211 on approval) | 4 | 2 | 3 | 9 | Medium | implementing |
 | 212 | Tech Debt | Replace hand-written `SafeFeature` / `GeoJSONFeature` with LinkML-generated equivalents — two hand-written TypeScript feature types in `shared/utils/src/types.ts` (one with a required, typed `geometry`; one with `geometry: SafeGeometry \| null` + `coordinates: unknown`). Pre-dates #200 but is an Article II tripwire — any "schema-adjacent" in-tree type should be LinkML-generated. #200 widens `calculateBounds` to a structural minimum that sidesteps the smell; this item closes it. (predates #200; follow-up) | 3 | 1 | 3 | 7 | Medium | proposed |
 | 211 | Tech Debt | Pre-computed-`bbox` fast-path in `@debrief/utils` `calculateBounds` — honour `feature.bbox` when present and skip the per-feature coordinate walk. `shared/components/src/utils/bounds.ts` already does this; `@debrief/utils` does not. Convergent-behaviour improvement that reduces one delta between the two implementations and is a prerequisite for any future unification (see #213). (follow-up to #200) | 3 | 1 | 4 | 8 | Low | proposed |
 | 210 | Tech Debt | Un-skip webview log-panel E2E suite — `tests/e2e/test-log-panel.spec.ts` is currently `test.describe.fixme(...)` pending issue #143 (webview iframe selector instability in code-server). Once #143 resolves, convert back to active tests; these cover the real integration path (code-server → LogPanel webview iframe → VS Code message bus) that Storybook cannot exercise. (requires #143, #176) | 3 | 1 | 4 | 8 | Low | proposed |
