+++ shared/components/src/StacBrowser/useBrowserFilter.ts
@@ -10,7 +10,7 @@
 import { useMemo, useRef } from 'react';
 import type { StacBrowserItem } from '../filter-engine/types';
 import type { ViewportPolygon, TimeFilter } from '@debrief/schemas';
-import { viewportToBounds, bboxOverlapsViewport } from '../utils/bounds';
+import { viewportToBounds, bboxOverlapsViewport } from '@debrief/utils';
 import { itemOverlapsFilter } from '../utils/timeline-helpers';
 import type { TemporalFilter } from '../TimelineView/types';
 import type { BrowserFilterResult } from './types';
