+++ shared/utils/src/bounds.ts
@@ -1,11 +1,34 @@
 /**
- * Bounds Calculation Utility
+ * Unified Bounds Utility Module — `@debrief/utils/bounds`
  *
- * Calculates bounding boxes from GeoJSON-like features for auto-zoom after
- * import and for "zoom to selection" on the VS Code map.
+ * Single canonical home for every bounds-related helper in the Debrief
+ * monorepo. Consolidates the previous `shared/components/src/utils/bounds.ts`
+ * copy into this module (feature 219, backlog #213).
+ *
+ * ## Supported feature-type families
+ *
+ * The input type for `calculateBounds` (and the other array-accepting helpers)
+ * is a structural minimum (`BoundsInputFeature`) that is satisfied without
+ * casts by all three external feature-type families used across the monorepo:
+ *
+ *   - `DebriefFeature` — LinkML-generated; from `@debrief/schemas`
+ *   - `SafeFeature`    — hand-written; `geometry: SafeGeometry | null`; from `@debrief/utils/types`
+ *   - `GeoJSONFeature` — raw JSON parse; from `@debrief/utils/types`
+ *
+ * This module does **not** re-export any of the above types — each caller
+ * imports its preferred family from its canonical location and passes it in.
+ * The structural minimum keeps this module decoupled from the LinkML
+ * `DebriefFeature` schema (Article II compliance).
+ *
+ * ## Public surface (9 functions)
+ *
+ *   calculateBounds, mergeBounds, boundsToLeaflet, isValidBounds  — existing
+ *   expandBounds, isPointInBounds, bboxOverlapsViewport,
+ *   viewportToBounds, filterBySpatialExtent                       — migrated from shared/components
  */
 
 import type { Bounds } from './types.js';
+import type { ViewportPolygon } from '@debrief/schemas';
 
 // Re-export Bounds type for convenience
 export type { Bounds };
@@ -19,9 +42,14 @@ export type { Bounds };
  * (`RawGeoJSONFeature`, `SafeFeature`, `DebriefFeature` and its variants) is
  * assignable to `ReadonlyArray<BoundsInputFeature>` via TypeScript's structural
  * subtyping — so no call site needs an `as`-cast.
+ *
+ * The optional `bbox` field supports the pre-computed-bbox fast-path (FR-008):
+ * when a feature carries a valid 4-number bbox tuple, `calculateBounds` uses
+ * it directly and skips the per-coordinate walk for that feature.
  */
 type BoundsInputFeature = {
   geometry?: { type: string; coordinates: unknown } | null | undefined;
+  bbox?: readonly number[] | null | undefined;
 };
 
 /**
@@ -34,6 +62,24 @@ type CoordinateTree =
   | number[][]        // LineString, MultiPoint
   | number[][][]      // Polygon, MultiLineString
   | number[][][][];   // MultiPolygon
+
+/**
+ * Private narrowing predicate for the pre-computed-bbox fast-path.
+ *
+ * Returns `true` iff `value` is a `Bounds`-shaped array: length ≥ 4, all four
+ * positional elements are finite numbers. Zero `any`, zero `as`-casts —
+ * the `value is Bounds` guard gives callers a typed view after checking.
+ */
+function isValidBboxTuple(value: unknown): value is Bounds {
+  return (
+    Array.isArray(value) &&
+    value.length >= 4 &&
+    Number.isFinite(value[0]) &&
+    Number.isFinite(value[1]) &&
+    Number.isFinite(value[2]) &&
+    Number.isFinite(value[3])
+  );
+}
 
 /**
  * Article XV.5 — explicit narrowing gate for untyped coordinate input.
@@ -112,6 +158,14 @@ function detectDepth(raw: unknown): 1 | 2 | 3 | 4 | null {
 /**
  * Calculate bounds from an array of GeoJSON-like features.
  *
+ * Accepts any of the three supported feature-type families without casts:
+ * `DebriefFeature[]`, `SafeFeature[]`, `GeoJSONFeature[]`, and the
+ * structural-minimum `BoundsInputFeature[]`.
+ *
+ * When a feature carries a valid pre-computed `feature.bbox`, the fast-path
+ * uses it directly and skips the per-coordinate walk for that feature, keeping
+ * map-fit latency O(n features) for STAC-style collections.
+ *
  * @param features Array of features with a structural `geometry` field.
  * @returns Bounds [minLon, minLat, maxLon, maxLat] or null if no valid coordinates.
  */
@@ -127,6 +181,19 @@ export function calculateBounds(
     if (!feature.geometry) {
       continue;
     }
+
+    // Pre-computed bbox fast-path (FR-008): honour feature.bbox when valid,
+    // skip coordinate walk for this feature. Falls back to coordinate walk for
+    // absent, null, or malformed bbox (FR-009).
+    if (feature.bbox !== undefined && feature.bbox !== null && isValidBboxTuple(feature.bbox)) {
+      const [fMinLon, fMinLat, fMaxLon, fMaxLat] = feature.bbox;
+      minLon = Math.min(minLon, fMinLon);
+      minLat = Math.min(minLat, fMinLat);
+      maxLon = Math.max(maxLon, fMaxLon);
+      maxLat = Math.max(maxLat, fMaxLat);
+      continue;
+    }
+
     const coords = coerceCoordinates(feature.geometry.coordinates);
     if (coords === null) {
       continue;
@@ -201,6 +268,128 @@ export function isValidBounds(bounds: Bounds): boolean {
   }
 
   return true;
+}
+
+/**
+ * Expand bounds by a percentage padding.
+ *
+ * @param bounds - Original bounds
+ * @param paddingPercent - Padding percentage (0.1 = 10%)
+ * @returns Expanded bounds
+ */
+export function expandBounds(bounds: Bounds, paddingPercent: number = 0.1): Bounds {
+  const [minLon, minLat, maxLon, maxLat] = bounds;
+  const lonRange = maxLon - minLon;
+  const latRange = maxLat - minLat;
+  const lonPad = lonRange * paddingPercent;
+  const latPad = latRange * paddingPercent;
+
+  return [
+    minLon - lonPad,
+    minLat - latPad,
+    maxLon + lonPad,
+    maxLat + latPad,
+  ];
+}
+
+/**
+ * Check if a point is within bounds.
+ */
+export function isPointInBounds(lon: number, lat: number, bounds: Bounds): boolean {
+  const [minLon, minLat, maxLon, maxLat] = bounds;
+  return lon >= minLon && lon <= maxLon && lat >= minLat && lat <= maxLat;
+}
+
+/**
+ * Check whether two axis-aligned bounding boxes overlap.
+ * Handles antimeridian crossing: when west > east the bbox is split
+ * into two halves for testing. west === east is treated as a zero-width
+ * bbox (degenerate point), NOT as an antimeridian crossing.
+ *
+ * @param itemBbox  - Item bounding box [west, south, east, north]
+ * @param viewportBbox - Viewport bounding box [west, south, east, north]
+ * @returns true if the two boxes overlap (including edge-touching)
+ */
+export function bboxOverlapsViewport(itemBbox: Bounds, viewportBbox: Bounds): boolean {
+  const [iW, iS, iE, iN] = itemBbox;
+  const [vW, vS, vE, vN] = viewportBbox;
+
+  // Latitude check — independent of antimeridian
+  if (iN < vS || iS > vN) return false;
+
+  const itemCrosses = iW > iE;
+  const vpCrosses = vW > vE;
+
+  if (!itemCrosses && !vpCrosses) {
+    // Standard AABB overlap on longitude
+    return !(iE < vW || iW > vE);
+  }
+
+  if (itemCrosses && !vpCrosses) {
+    // Item crosses antimeridian → split item into [iW, 180] and [-180, iE]
+    return !(vE < iW && vW > iE);
+  }
+
+  if (!itemCrosses && vpCrosses) {
+    // Viewport crosses antimeridian → split viewport
+    return !(iE < vW && iW > vE);
+  }
+
+  // Both cross antimeridian — they always overlap longitudinally
+  return true;
+}
+
+/**
+ * Convert a ViewportPolygon (4-corner [NW, NE, SE, SW]) to an axis-aligned Bounds.
+ * For non-rotated views, this extracts [minLon, minLat, maxLon, maxLat].
+ * For rotated views, this computes the enclosing AABB.
+ *
+ * Returns null for degenerate polygons (zero area).
+ * Feature: 132-three-view-sync, updated: 203 (object-form Coordinate).
+ *
+ * @remarks
+ * This function is specific to 4-corner ViewportPolygon inputs. It uses
+ * `Math.min(...lons)` / `Math.max(...lons)` which collapse to spread arguments —
+ * V8 rejects spreads with more than ~100k arguments, so do NOT reuse this on
+ * large coordinate arrays (FR-022). For large arrays, replace the spread with
+ * a for-loop accumulator.
+ *
+ * @param viewport - 4-corner polygon [NW, NE, SE, SW] with `{ longitude, latitude }` corners.
+ * @returns Bounds tuple [minLon, minLat, maxLon, maxLat] or null if degenerate
+ */
+export function viewportToBounds(viewport: ViewportPolygon): Bounds | null {
+  const coords = viewport.coordinates;
+  const lons = coords.map((c) => c.longitude);
+  const lats = coords.map((c) => c.latitude);
+
+  const minLon = Math.min(...lons);
+  const maxLon = Math.max(...lons);
+  const minLat = Math.min(...lats);
+  const maxLat = Math.max(...lats);
+
+  // Degenerate polygon — zero area (all corners at same point or on a line)
+  if (minLon === maxLon || minLat === maxLat) {
+    return null;
+  }
+
+  return [minLon, minLat, maxLon, maxLat];
+}
+
+/**
+ * Filter items to those whose bbox overlaps the given viewport.
+ * Items without a bbox are excluded.
+ *
+ * @param items - Array of items with an optional bbox property
+ * @param viewportBbox - Current viewport bounds
+ * @returns Filtered array of items whose bbox overlaps the viewport
+ */
+export function filterBySpatialExtent<T extends { bbox: Bounds | null }>(
+  items: readonly T[],
+  viewportBbox: Bounds,
+): T[] {
+  return items.filter(
+    (item): item is T => item.bbox !== null && bboxOverlapsViewport(item.bbox, viewportBbox),
+  );
 }
 
 /**
